GustyDAG: Create a DAG in one line of code based on METADATA.yml
Introduces the gusty.GustyDAG class, which initializes a DAG given a folder. The folder may contain a METADATA.yml file that specifies arguments to the DAG, such as schedule_interval or default_args. This would simplify each DAG .py file (could be just a couple lines of code) and move a lot of the logic to the folder.

Also introduces two other things:

* GustyYAMLLoader is a YAML loader with a few helpful Airflow-specific functions. For example, in a YAML file you can represent days_ago with `start_date: !days_ago 1`. Especially helpful for the metadata loading, but I've changed the operator parsing to use this as well.

* examples/gusty_tutorial, which copies the default Airflow tutorial in the Gusty style.

This is a flexible pattern that could keep being extended in the future; often the same default arguments will be shared across many DAGs in Airflow, so one could imagine introducing something like a default METADATA.yml in the dags folder.